+++ client/src/modals/RankingModal/Rankbar/index.style.tsx
