+++ client/src/modals/RankingModal/Selector/index.style.tsx
