+++ client/src/modals/RankingModal/index.style.tsx
