--- client/src/pages/RankingPage.tsx
@@ -1,9 +1,0 @@
-import RankingModal from '@components/RankingModal';
-
-import React from 'react';
-
-const RankingPage: React.FC = () => {
-  return <RankingModal />;
-};
-
-export default RankingPage;
