--- client/src/pages/ReviewSubmitPage.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import ReviewModal from '@components/ReviewModal';
-
-const ReviewSubmitPage: React.FC = () => {
-  return <ReviewModal />;
-};
-
-export default ReviewSubmitPage;
