--- client/src/pages/SignInPage.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import SignInModal from '@components/SignInModal';
-
-const SignInPage: React.FC = () => {
-  return <SignInModal />;
-};
-
-export default SignInPage;
