--- client/src/pages/SignUpPage.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import SignUpModal from '@components/SignUpModal';
-
-const SignUpPage: React.FC = () => {
-  return <SignUpModal />;
-};
-
-export default SignUpPage;
