+++ client/src/pages/index.ts
@@ -1,23 +1,7 @@
 import NotFoundPage from '@pages/NotFoundPage';
 import MainPage from '@pages/MainPage';
-import ReviewSubmitPage from '@pages/ReviewSubmitPage';
-import RankingPage from '@pages/RankingPage';
-import SignInPage from '@pages/SignInPage';
 import LoadingPage from '@pages/LoadingPage';
-import SignUpPage from '@pages/SignUpPage';
 import ProfilePage from '@pages/ProfilePage';
-import ProfileAddressPage from '@pages/ProfileAddressPage';
 import LoadPage from '@pages/LoadPage';
 
-export {
-  NotFoundPage,
-  MainPage,
-  ReviewSubmitPage,
-  RankingPage,
-  SignInPage,
-  LoadingPage,
-  LoadPage,
-  SignUpPage,
-  ProfilePage,
-  ProfileAddressPage,
-};
+export { NotFoundPage, MainPage, LoadingPage, LoadPage, ProfilePage };
